A lot Upgrade

+++ GuardApp/GuardPersonalForm.cs
@@ -8,7 +8,9 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using GuardApp.Model;
+using GuardApp.Model.HelperModel;
 using GuardApp.Repository;
+using GuardApp.Helper;
 
 namespace GuardApp
 {
@@ -20,7 +22,7 @@ namespace GuardApp
             this.BackgroundImage = Properties.Resources.tr;
             BackgroundImageLayout = ImageLayout.Stretch;
         }
-        // liste olayını degıstır.
+
         Repository<GuardPersonal> guardPersonalRepository = new Repository<GuardPersonal>();
         Repository<Personal> personalRepository = new Repository<Personal>();
         Repository<Guard> guardRepository = new Repository<Guard>();
@@ -28,11 +30,13 @@ namespace GuardApp
         int relatedGuardId = 0;
         List<Personal> relatedPersonals = new List<Personal>();
         List<Personal> otherPersonals = new List<Personal>();
+        List<PersonalViewModal> otherChangedList = new List<PersonalViewModal>();
         private void GuardPersonalForm_Load(object sender, EventArgs e)
         {
             FillGuardListBox();
             FillPersonalListBoxes();
-            // USTTEK METHODLARI DUZENLE . SAYFA YUKLENDIGINDE OTOMATIK VAR OLAN KAYITLARI GETIR DEGISTIRILDIGINDE ISE GUNCELLE         
+            // USTTEK METHODLARI DUZENLE . SAYFA YUKLENDIGINDE OTOMATIK VAR OLAN KAYITLARI GETIR DEGISTIRILDIGINDE ISE GUNCELLE
+            ShowGuardNameOnLabel();
         }
 
         private void btnBack_Click(object sender, EventArgs e)
@@ -51,28 +55,48 @@ namespace GuardApp
 
         private void FillRelatedPersonalListBox()
         {
-            var relatedPersonalId = guardPersonalRepository.List().Where(x => x.GuardId == relatedGuardId&&x.Personal.IsActive==true).ToList().Select(x => x.PersonalId).ToList();
-
-             relatedPersonals = personalRepository.List().Where(x => relatedPersonalId.Contains(x.Id)&&x.IsActive==true).ToList();
-
-            lstGuardPersonal.DataSource = null;
-            if (relatedPersonals != null)
+            try
             {
-                lstGuardPersonal.DataSource = relatedPersonals.ToList();
-                lstGuardPersonal.ValueMember = "Id";
-                lstGuardPersonal.DisplayMember = "Name";
+                var relatedPersonalId = guardPersonalRepository.List().Where(x => x.GuardId == relatedGuardId && x.Personal.IsActive == true).ToList().Select(x => x.PersonalId).ToList();
+                relatedPersonals = personalRepository.List().Where(x => relatedPersonalId.Contains(x.Id) && x.IsActive == true).ToList();
+
+                List<PersonalViewModal> personalViewModals = relatedPersonals.PersonalDisplayerFormatList();
+
+                lstGuardPersonal.DataSource = null;
+                if (personalViewModals != null)
+                {
+                    lstGuardPersonal.DataSource = personalViewModals.ToList();
+                    lstGuardPersonal.ValueMember = "PersonalId";
+                    lstGuardPersonal.DisplayMember = "Name";
+                }
             }
+            catch (Exception)
+            {
+
+                throw;
+            }
+
         }
 
         private void FillAllPersonalListBox()
         {
-            otherPersonals = personalRepository.List().Where(x=>x.IsActive==true).Except(relatedPersonals).ToList();
-            lstAllPersonal.DataSource = null;
-            if (otherPersonals != null)
+            try
             {
-                lstAllPersonal.DataSource = otherPersonals.ToList();
-                lstAllPersonal.ValueMember = "Id";
-                lstAllPersonal.DisplayMember = "Name";
+                otherPersonals = personalRepository.List().Where(x => x.IsActive == true).Except(relatedPersonals).ToList();
+
+                otherChangedList = otherPersonals.PersonalDisplayerFormatList();
+                lstAllPersonal.DataSource = null;
+                if (otherChangedList != null)
+                {
+                    lstAllPersonal.DataSource = otherChangedList.ToList();
+                    lstAllPersonal.ValueMember = "PersonalId";
+                    lstAllPersonal.DisplayMember = "Name";
+                }
+            }
+            catch (Exception)
+            {
+
+                throw;
             }
         }
 
@@ -86,7 +110,7 @@ namespace GuardApp
             {
                 lstAllPersonal.DataSource = null;
                 lstAllPersonal.Items.Clear();
-                foreach (var personal in otherPersonals)
+                foreach (var personal in otherChangedList)
                 {
                     if (personal.Name.ToUpper().Contains(searchBox.Text.ToUpper()))
                         lstAllPersonal.Items.Add(personal);
@@ -96,46 +120,70 @@ namespace GuardApp
         //Tüm personellerden seçilen personeli , Nöbete ata buton
         private void btnAddPersonalToGuard_Click(object sender, EventArgs e)
         {
-            var relatedPersonal = (Personal)lstAllPersonal.SelectedItem;
-            guardPersonalRepository.Insert(new GuardPersonal() { GuardId = relatedGuardId, PersonalId = relatedPersonal.Id });
-            FillPersonalListBoxes();
+            var relatedPersonalViewModal = (PersonalViewModal)lstAllPersonal.SelectedItem;
+            if (relatedPersonalViewModal != null)
+            {
+                guardPersonalRepository.Insert(new GuardPersonal() { GuardId = relatedGuardId, PersonalId = relatedPersonalViewModal.PersonalId });
+                FillPersonalListBoxes();
+            }
         }
         //Tüm personellerden seçilen personeli , Nöbete ata çift tıklandığında
         private void lstAllPersonal_DoubleClick(object sender, EventArgs e)
         {
-            var relatedPersonal = (Personal)lstAllPersonal.SelectedItem;
-            guardPersonalRepository.Insert(new GuardPersonal() { GuardId = relatedGuardId, PersonalId = relatedPersonal.Id });
-            FillPersonalListBoxes();
+            var relatedPersonalViewModal = (PersonalViewModal)lstAllPersonal.SelectedItem;
+            if (relatedPersonalViewModal != null)
+            {
+                guardPersonalRepository.Insert(new GuardPersonal() { GuardId = relatedGuardId, PersonalId = relatedPersonalViewModal.PersonalId });
+                FillPersonalListBoxes();
+            }
         }
         //Nöbetteki personeli , Nöbetten çıkar buton
         private void btnPersonalRemove_Click(object sender, EventArgs e)
         {
-            var relatedPersonal = (Personal)lstGuardPersonal.SelectedItem;
-            var relatedPersonalGuard = guardPersonalRepository.List().FirstOrDefault(x => x.GuardId == relatedGuardId && x.PersonalId == relatedPersonal.Id);
-            guardPersonalRepository.Delete(relatedPersonalGuard);
-            FillPersonalListBoxes();
+            var relatedPersonalViewModal = (PersonalViewModal)lstGuardPersonal.SelectedItem;
+            if (relatedPersonalViewModal != null)
+            {
+                var relatedPersonalGuard = guardPersonalRepository.List().FirstOrDefault(x => x.GuardId == relatedGuardId && x.PersonalId == relatedPersonalViewModal.PersonalId);
+                guardPersonalRepository.Delete(relatedPersonalGuard);
+                FillPersonalListBoxes();
+            }
+
         }
         //Nöbetteki personeli , Nöbetten çıkar  doubleclick
         private void lstGuardPersonal_DoubleClick(object sender, EventArgs e)
         {
-            var relatedPersonal = (Personal)lstGuardPersonal.SelectedItem;
-            var relatedPersonalGuard = guardPersonalRepository.List().FirstOrDefault(x => x.GuardId == relatedGuardId && x.PersonalId == relatedPersonal.Id);
-            guardPersonalRepository.Delete(relatedPersonalGuard);
-            FillPersonalListBoxes();
+            var relatedPersonalViewModal = (PersonalViewModal)lstGuardPersonal.SelectedItem;
+            if (relatedPersonalViewModal != null)
+            {
+                var relatedPersonalGuard = guardPersonalRepository.List().FirstOrDefault(x => x.GuardId == relatedGuardId && x.PersonalId == relatedPersonalViewModal.PersonalId);
+                guardPersonalRepository.Delete(relatedPersonalGuard);
+                FillPersonalListBoxes();
+            }
         }
 
         private void lstGuard_MouseClick(object sender, MouseEventArgs e)
         {
+            ShowGuardNameOnLabel();
             FillPersonalListBoxes();
         }
 
         private void FillPersonalListBoxes()
         {
             var selectedGuard = (Guard)lstGuard.SelectedItem;
-            if (selectedGuard!=null)            
-                relatedGuardId = selectedGuard.Id;           
+            if (selectedGuard != null)
+                relatedGuardId = selectedGuard.Id;
             FillRelatedPersonalListBox();
             FillAllPersonalListBox();
-        }   
+        }
+
+        private void ShowGuardNameOnLabel()
+        {
+            Guard selectedGuard = (Guard)lstGuard.SelectedItem;
+            if (selectedGuard != null)
+            {
+                lblDisplayGuard.Text = selectedGuard.Name;
+            }
+        }
+
     }
 }
